Buy Button Update
Buy button doesn't work as I need to make the shop side bar first.
In the shop sidebar, I will have the name of the item, price of the item and total cost of all those items
At the very bottom, it will show the total cost of the total of EVERY item.